+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Data Transfer Object classes generation from swagger spec
+
+## [0.0.10] - 2020-07-19
+### Added
 - AbstractDocument::fromHttpMessage(), possible constructing DTO from Response and Requests both
 - PsrContainerAwareSchemaContainer for injection dependencies to neomerx Schemas
 
@@ -79,8 +82,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - FastRoute Dispatcher generation from swagger operationIds
 
-[Unreleased]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.9...HEAD
-[0.0.8]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.8...0.0.9
+[Unreleased]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.10...HEAD
+[0.0.10]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.9...0.0.10
+[0.0.9]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.8...0.0.9
 [0.0.8]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.7...0.0.8
 [0.0.7]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.6...0.0.7
 [0.0.6]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.5...0.0.6